refactor: db 전용 dto 패키지 service 패키지 안으로 이동
