+++ src/main/java/racingcar/service/GameService.java
@@ -9,8 +9,8 @@ import racingcar.domain.Lap;
 import racingcar.domain.NumberGenerator;
 import racingcar.domain.RandomNumberGenerator;
 import racingcar.domain.WinnerMaker;
-import racingcar.dto.request.GameRequestDto;
-import racingcar.dto.response.GameResponseDto;
+import racingcar.service.dto.GameRequestDto;
+import racingcar.service.dto.GameResponseDto;
 import racingcar.entity.Game;
 import racingcar.entity.PlayerResult;
 import racingcar.repository.GameRepository;
