+++ src/test/java/racingcar/service/GameServiceTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import racingcar.dto.request.GameRequestDto;
-import racingcar.dto.response.GameResponseDto;
+import racingcar.service.dto.GameRequestDto;
+import racingcar.service.dto.GameResponseDto;
 import racingcar.entity.Game;
 import racingcar.entity.PlayerResult;
 import racingcar.repository.GameRepository;
